moved mocks to tests folder
